move all moduleFallback logic into its own file
